Add support for internal synchronization of unit list fields.
This patch adds support for internal synchronization of unit list
fields, i.e., list fields annotated with a `&synchronize` attribute will
emit code which if during parsing the list the parser sees input which
cannot be parsed into a list element (e.g., due to additional data in
the input, or data bein missing), we will enter synchronization mode.

List fields currently come in two relevant flavors: lists using
lookahead (e.g., lists with no explicit size, but determining their stop
condition dynamically), and lists with a fixed size. For lists with
lookahead we will attempt to find a new list element, while for
fixed-size lists we skip the faulty element and continue with the next
one.

This feature is orthogonal to using a list field as a synchronization
point for previous fields which we already supported.