Make `confirm` and `reject` builtin statements.
We previously modelled `confirm` and `reject` as unit methods. This
worked well enough in a world where a unit always has to leave trial
mode to parse successfully.

We removed that restriction since then and made trial mode a global
property. In order to mutate the global trial mode `confirm` and
`reject` need to be invoked on the currently parsing unit. While
invoking them on other units was possible, it had no effect. This was
error prone.

With this patch we make `confirm` and `reject` freestanding statements.
Under the covers this still invokes the unit-specific logic but always
on the current unit.