Make sure user-visible enums are represented by `int64_t`.
In order to convert from arbitrary runtime values with
`hilti::rt::enum_::{from_int, from_uint}` we statically assert that the
enums have an underlying type `int64_t`. This is not true in general,
e.g., a port enum would previously have been represented by an `int`
(which not only does not need to be the same time, but could also have a
smaller range).

With this path we now explicitly request an underlying type `int64_t`
for enums which can be constructed by users from Spicy code.

Closes #1122.