Optimize out unused hooks related to synchronization.
The free-standing functions `spicy_rt::confirm` and `spicy_rt::reject`
are implemented in C++ and call parser hooks. This means that in order
to track feature requirements for these functions we need to use
function parameter attributes on the free-standing functions.

Now if the hooks are not implemented they would previously have been
removed, even if the hooks would have been annotated with feature
requirements. This was due to the optimizer not taking feature
requirements into account when collecting field information.

With this patch this part of the optimizer now also takes feature
requirements into account. This allows us to move to feature requirement
flags for hooks related to synchronization once we also annotate the
free-standing functions calling the hooks. We can then drop the
`&always-emit` attribute these hooks previously had which allows us to
elide them if they are unused.

Closes #1131.