Fix constness for dereferencing references.
We don't want the constness of the reference type transfer over to
the result.

Note: I think this is the right change, but it comes with one piece of
oddity, as demonstrated by the included test baseline update:

    type Foo = unit(x: bool, y: Bar&)

`y` now becomes non-const, because behind the scences this gets mapped
to `strong_ref<Bar>`, and it's the reference that becomes constant due
to being an `in` parameter, not the inner type. I think this is ok,
although it does turn the `y` essentially into an `inout` parameter.

Closes #1112.