Support unit initialization through a struct constructor expression.
This works now:

        type X = unit {
            a: uint8;
            b: uint8;
        };

        ...

        local x: X = [$a = 1, $b = 2];

We implement this by first coercing the RHS struct ctor into a unit
ctor; and then at codegen time turn it back into an equivalent struct
ctor. That's a bit circly, but necessary to get the Spicy-side type
checking right. Validation is left to the final HILTI struct
expression.

Closes #1036.