Fix clang-tidy `HeaderFilterRegex`.
In general LLVM's regexp engine does not seem to support negative
matches at all. We previously tried to set this up, but this also
filtered many headers we would have liked to lint.

This patch sets a regexp under the assumptions that our paths do not
contain numbers. With that many more diagnostics are triggered which we
also fix.