Adjust signature in test.
This test checks that a declaration in HILTI can find a function defined
in C++ code, but the signatures of the generated HILTI declaration and
the actual definition did not match (HILTI declared return type
`integer::safe<int64_t>`, but defined C++ function returned `int64_t`).
Astonishingly, this test still linked and ran successfully on 64bit
platforms. On 32bit platforms on the other hand the external function
seems to return random data.

This patch harmonizes the return types so the test becomes
deterministic.