Fix linking of runtime unit test executable for static build.
Depending on whether we link Spicy statically or dynamically, we switch
between linking a shared library or a CMake object library. The runtime
unit test executables previously never used the object libraries, but
would in the static linking case link against an archive file. This
could lead to undefined references.

With this patch we use the existing helper macros to ensure correct
linking for the runtime unit test executable as well.