Remove freebsd11, add freebsd13.
Freebsd11 is EOL and it seems upstream package servers became
unavailable. This makes it impossible to install required packages. This
patch adds support for freebsd13 instead. Since it seems that support the
the way we perform symbol demangling is currently broken in freebsd13
(see e.g., https://bugs.freebsd.org/bugzilla/show_bug.cgi?id=222563)
the step is not required.

This patch is the first, but not last piece of work to get #1171 fixed.