Avoid truncation when performing int64 range check.
While in principle `std::abs` should have an overload taking a `long
long` so that passing an `int64_t` work as expected, this overload seems
to be missing e.g., on freebsd-13. Similarly `std::labs` and
`std::llabs` seem to be also operating with an `int` argument.

This path replaces uses of `std::abs` when performing an int64 range
check during parsing with `std::fabs` to avoid truncation issues with
non-compliant `std::abs` implementations.