Fix linking of runtime test executables.
We previously directly linked to the actual runtime libraries instead of
the corresponding object libraries. On some platforms this could have
lead to unresolved symbols when building with static libraries.

With this patch we now link against the object libraries (for the actual
objects) and their corresponding actual libraries as well (for include
paths which are not properly propagated for object libraries).