Only check whether Zeek docs are up-to-date from cron builds.
The check does not and as currently written conceptually cannot use
pinned versions of the external repos. This causes issues where the
state of generated docs and the external dependencies gets out of sync
over time (e.g., on older release branches), so this patch runs the
check only when executed in a cron build. Since the only cron build we
run currently is on for the `main` branch we can still check whether
docs get out of sync there.