Cap parallelism use for JIT background jobs.
Compilation during JIT happens on background jobs spawned in the
background. We previously would spawn all required jobs in parallel
which for projects making heavy use of modules lead to a huge number of
background jobs which were all competing for CPU and memory (the latter
being much more severe as in case especially memory-hungry compilers were
used this could lead to jobs being OOM killed). The only way out for
users was to set `HILTI_JIT_SEQUENTIAL` which would remove _all
parallelism_ which for module heavy projects would have caused very long
compile times.

In this patch we introduce new infrastructure so the maximum number of
background jobs can be capped. For that we separate the notions of
scheduling and actually spawning a job. This allows us to make
background jobs "lazy" in that they are only actually spawned if waited
for. Practically, instead of immediately spawning jobs, JIT now just
schedules them; when then waiting for the jobs we spawn parallel jobs up
the set parallelism.

By default we spawn up to `nproc` parallel jobs. We introduce the
environment variable `HILTI_JIT_PARALLELISM` which can hold an unsigned
integer giving the maximum number of parallel process to use instead.
With that `HILTI_JIT_SEQUENTIAL` becomes equivalent to
`HILTI_JIT_PARALLELISM=1`, but we keep support for
`HILTI_JIT_SEQUENTIAL`.

Closes #1179.