Add builder method to set a new location.
This patch adds a builder method which set a new location  which will be
used e.g., for runtime errors if no other location has been set in the
meantime. For that we introduce a new AST node which captures that
location information so it can be rendered to HILTI (currently as a
comment), and to generate dedicated C++ code which updates the location.
Since the C++ code just sets a pointer to a `const char*` we ATM cannot
expose this as a pure HILTI function since HILTI has no concept of
non-owning (or even static duration) strings.

We will use this in a subsequent patch.