Fix support for having multiple source modules of the same name.
There was an underlying assumption in much of the code that a module
name would be globally unique, meaning it could be used as index into
maps and as seeds for C++ identifiers. However, while that's often the
case, the `import ... from ...` syntax allow pulling in multiple
modules all having the same name.

This change fixes that to track a module's "scope" (i.e., the part
after `from ...`) more thoroughly across various pieces. In
particular, we now expose a new `hilti::Unit::uniqueID()` method that
incorporates that scope to provide a globally unique identifier for
each module (if there's no scope, `uniqueID()` will return the same as
`id()`, otherwise it will include a normalized version of the scope).

This new `uniqueID()` is now used at various place that require unit
IDs to be globally unique; and also in debug output for easier
tracking of multiple modules having the same name.

This all came up with #1187, which indeed uses `import ... from ...`
to import two different modules both of the name `foo`. Without this
fix, our module cache would return the first unit instance for the
second `import` statement because the matching name suggested that the
module was already loaded.

Closes #1187.