Make handling of sanitizer workarounds more granular.
We previously did not distinguish between which sanitizers where enabled
on the source level; we also did not detect whether an analyzer was
disabled directly via e.g., `CMAKE_CXX_FLAGS` instead of via our
configure flag. This could have caused issues when enabling sanitizers
other than ASAN, or when embedding Spicy as a subdirectory in another
project.

With this patch we now directly detect whether we need to execute
sanitizer-specific code (for now only needed for ASAN).