Disable building of benchmark target by default.
Benchmark's configure step seems to sometimes run into detect the regexp
backend (`Failed to determine the source files for the regular
expression backend`). This patch changes our CMake setup to only include
the benchmark targets if explicitly requested. In order to at least
sometimes exercise this code we also change `ci/run-ci` so that
benchmarks are by default enabled when using that script (but could also
be disabled from there with a configure option).