Adjust removal of symlinks on install for `DESTDIR`.
We previously did not take `DESTDIR` installs into account when removing
symlinks in the installation directory, so we never actually removed
symlinks for `DESTDIR` installs. With this patch we now check whether a
`DESTDIR` is set in the environment an optionally prefix installation
paths with it when cleaning up symlinks (`DESTDIR` will already end in a
`/`).

Closes #1201.