Remove potential use-after-move.
Since in C++ the evaluation order of arguments is unspecified, accessing
members of an argument while also moving it in the same argument list
could potentially access moved-from data. This patch introduces a
temporary to remove such a potential bug which we actually see break
even without optimizations with gcc-12.1.

Closes #1204.