Prevent unnecessarily executable stack with GNU toolchain.
The GNU toolchain seems to default to marking the stack of any archive
as executable if it contains assembly sources. This is the case in Spicy
since we link against 3rdparty/fiber which contains assembly source
files. If Spicy is linked statically this propagates into the binary,
for shared linkage only into libhilti; in each case this can be
confirmed with e.g., `readelf -Wl <artifact>`.

With this patch we explicitly request that the stack not to be marked
executable whenever we link fiber assembly files. Since the library does
not provide a CMake customization point for the flags we set the flag
globally for any 3rdparty sources.

Closes #1210.