Fix links to Fedora artifacts for releases.
We dropped support for Fedora 33 and 34 after the last release which
means there are no more binary packages from a dev snapshot for them.
For the added Fedora 35 and 36 on the other hand, only packages for dev
snapshots exist, but we add links to the artifacts to the coming release
(not yet available).