Fix code generation when a module's file could be imported through different means.
This can happen when mixing normal import with  `import ... from`. A
previous change to support multiple units of the same name, missed
that we could run into namespacing trouble at code generation time in
this case. The problem was that we included the `from ...` scope into
the namespace, which doesn't work when we import the same module both
w/ and wo/ scope. Now we generate unique module IDs differently
through global state independent of file system paths.

Close #1227.