Add special handling for potential `advance` failure in trial mode.
During trial mode we might call `advance` to find a synchronization
point. If any of the bits touched by that fall into a gap, this can
trigger a `MissingData` exception.

In this patch we emit an additional `try`/`catch` block around the
`advance` if we search for a lookahead in `Try` mode. We can then
recover from `MissingData` by jumping to the next non-gap block in the
input.

Closes #1231.