When decoding bytes into a string using a given charset, allow caller to control error handling.
All methods taking a charset parameters now take an additional enum
selecting 1 of 3 possible error handling strategies in case a
character can't be decoded/represented: `STRICT` throws an error,
`IGNORE` skips the problematic character and proceeds with the next,
and `REPLACE` replaces the problematic character with a safe
substitute. `REPLACE` is the default everywhere now, so that by
default no errors are triggered.

This comes with an additional functional change for the ASCII
encoding: we now consistently sanitize characters that ASCII can't
represent when in `REPLACE`/`IGNORE` modes (and, hence, by default),
and trigger errors in `STRICT` mode. Previously, we'd sometimes let
them through, and never triggered any errors. This also fixes a bug
with the ASCII encoding sometimes turning a non-printable character
into multiple repeated substitutes (which our unit tests actually then
checked for; fixed the tests, too).

Closes #1251.