Redo attribute coercion in Spicy.
For a couple of field attributes, we performed coercion of their
values during parser generation time, which isn't where this should be
done (because it won't be reflected in the finalized AST in that
case). Moving this into a new Spicy-side coercer pass (which was still
missing).

We should probably add coercion for more attributes there, but the existing
code doesn't have that either, so it seems not necessary (so far ...).

(No new tests as existing tests cover this.)