Fix contexts not allowing being passed `inout`.
This turned into a larger fix needing some internal changes, but it's
a good change because it unifies a couple of hacks we had so far, and
probably also prevents some other trouble we hadn't even found yet.

The problem was that `inout` parameters require a LHS value that can
be passed by reference, but our C++ code generator had no way of
telling if a C++-side expression ended up representing a RHS or LHS
value; and hence couldn't turn a RHS into a LHS when it was needed for
passing by reference.

We had a couple of places in the code generator already were we
accounted for this manually, but generally the problem wasn't solvable
without consistency tracking the LHS/RHS information. This change now
adds  that state to `cxx::Expression` so that it can be set when the
expression is created and then follows it around to where it's needed.