Embed a C function into hilti-rt that encodes our version in its name.
If a host applications links against this, it will fail to load if the
runtime version changes without a rebuild. This is the same approach Zeek
is using to ensure plugins are loaded only into the Zeek version they
were compiled for.

Addresses #1221.
Addresses https://github.com/zeek/spicy-plugin/issues/82