Tie HLTOs files to the Spicy version that generated them.
They will now fail at load time if the current Spicy version doesn't
match what they were built with.

Not quite sure if this should obsolete our current
`__hlt_library_version` information, which is a runtime version check
that so far only *warned* about mismatches. As that information is
part of the `open()` API for HLTOs, and stores some additional
information as well, I've kept it for now.