Replace `Lambda` class.
We originally introduced `Lambda` to reduce compilation overhead from
use of `std::function`, but the code we pulled in was under a license
which is incompatible with the licenses typically used for other Zeek
projects.

Since we only ever used `Lambda` to store exactly one type of callback
we now replace it with a much smaller custom implementation. We still
require type-erase to store lambdas, but know both argument as well as
return types.

Closes #1273.