Fix duplicate prefixing of `::` for prototypes generated for `&cxxname`.
If the identifier used for the `&cxxname` of a function already started
with `::` we would previously always have prefixed with `::`. When
emitting a declaration in the namespace we would strip a leading `::` in
the emitted namespace, but would potentially have left a stray `::` at
the start which is invalid C++ syntax.

We now only inject a leading `::` in the identifier if the name is not
prefixed already.