Support Zeek-style documentation strings in Spicy source code.
This adds support for extracting and retaining Zeek-style
documentation strings (`##`, `##!`) inside Spicy source code to any
top-level AST nodes where it's appropriate: if they are prefixed with
a block of such comments, that information is retained along with the
node for later access.

This commit also prepares for supporting `##<` to document unit and
struct fields, but that isn't implemented yet because it will need
another refactor of the Bison grammar.

The code printer (`-p`) learned to emit doc strings coming with the
nodes it's printing out. Beyond that, we do not make further use of
them yet, but may later leverage them for auto-generating Spicy-side
documentation, which we currently do through a Python script ingesting
source code directly.