Support Zeek-style documentation strings in Spicy source code.
This adds support for extracting and retaining Zeek-style
documentation strings (`##`, `##!`) in Spicy source code to all
declarations: if they are prefixed with blocks of such comments, that
information is retained along with the declaration for later access.

This commit also prepares for supporting `##<` to document unit and
struct fields, but that isn't implemented yet.

We don't actually use the extracted information further yet, but may
later leverage it for auto-generating Spicy-side documentation (which
we currently do through a Python script ingesting source code
directly.)

Originally, this was implemented to support the Zeek plugin providing
documentation to Zeekygen, but we ended up using a different approach
there (prototyping events Zeek-side). Still, seems worth keeping this
capability.