Fix incorrect offset computation in `advanceToNextData`.
To compute the next `View` in `advanceToNextData` after a gap chunk we
would previously compute offsets incorrectly in that we would implicitly
assume that the begin of the original `View` started with zero offset.
This is unlikely true for any but the most trivial cases and often caused e.g.,
resynchronization to jump to incorrect offsets much after the correct
position.

With this patch we now use a correct offset computation.

Closes #1303.