Expose `pack`/`unpack` to Spicy.
So far both were available at the HILTI-level only. We now make them
available in Spicy, too. However, in Spicy we change error semantics
so that `unpack` throws an exception if it can't parse the data,
instead of returning a `result` type (which (1) we aren't using in Spicy
yet, and (2) wouldn't match with how errors are treated elsewhere
anyways).