Extend `unpack` operator to optionally throw an exception on error.
This internally adds a constant boolean parameter to the `unpack`
operator that, if true, lets it throw an `InvalidValue` exception if
it can't parse its input, instead of returning a `result`. In that
case the return type of the operator also changes to now directly
return the 2-tuple that would otherwise be contained inside the
`result`.

This is an strictly internal change for now, the HILTI language
continues to always return a `result` by hardcoding the boolean to
`false`. We will later use this to change the error semantics for the
Spicy version of the operator, which a subsequent commit will
introduce.

Because there's no visible effect of this internal change currently,
there's also no test. We will test it at the Spicy-level once `unpack`
is introduced there.