Support arbitrary expression as argument to `interval(...)` and `interval_ns(...)`.
So far we used constructor expressions like `interval(...)` to create
*constants* of the given type, and for that we required that the
argument was a single atomic value. The result was that these
constructor expressions were really more like literals for the
corresponding types than "normal" expressions. While that's useful
internally, it's confusing for the user. This commits is beginning to
change that, using `interval`/`interval_NS` as the first instances
that now supports arbitrary expressions as its argument. In follow-up
commits, we'll port other constructor expressions over to that new
syntax. (This also unifies HILTI & Spicy to each provide both of those
interval constructors.)

To make this change without loosing a way to create actual constants
(which we need at some places), we implement this in two stages: the
parser initially turns such `interval(...)` expressions into call
operators[1] that the types now define for each desired constructor. We
then extend the normalizer to look at all the relevant call operators:
if they are just receiving a constant value as their argument, then it
substitutes them with a plain ``Ctor`` node (``ctor::Interval`` in our
cases here).

That check for "if they just receive a constant value" isn't quite
trivial in general. In our case, where we just want to go back to what
we had, it's not that difficult, but more generally it would be nice
if we had a constant folder that could compute results even for more
complex, but constant, expressions. As that's a piece that our AST
infrastructure is still missing, this commit starts going that route
by adding a new "constant folder" visitor. Right now, it's covers only
the simple cases we need here, but it's something we can extend going
forward.

Addresses #1123.

[1] These call operators are a bit special because they actually don't
have anything to call: in `interval(42)` the `interval` isn't any kind
of actual value; we just need it for overload resolution. To make that
work, we use a `expression::Member` as the first operand, which models
a predefined ID. Normally `Member` is just for field lookups in, e.g.,
structs, but it works just as well elsewhere.