Validate that unit using `self.context()` declares `%context`.
When declaring `context` methods we previously would `assert` that any
unit using `self.context()` actually would declare a `%context`. This is
not true and lead to segfaults in release builds which do not execute
code in `assert`.

This patch adds new validation which rejects uses of `self.context()` in
units which did not declare a context. Since validation happens after
declaration of the `context` methods has executed (the part which
previously had the `assert`), we return a dummy type (a void) for such
units, and then reject the code later.

Closes #1311.