Improve efficiency of `startsWith` for long inputs.
We previously would always search for _any occurrence_ of `prefix` in the
given string and only later check whether it was at the beginning; this
could have meant potentially lots of work for long input strings.

With this patch we examine at most the first `prefix.size()` characters
of the string.