Bring implementation of `Sink::write` in line with declaration.
In `spicy_rt.hlt` `Sink::write` is declared as taking an `in` `data`
parameter while the C++ implementation declared a mutable parameter,
i.e., `inout`. Surprisingly, this resolved fine most of the time, but
failed to resolve when actually passing a `const` value.

This patch brings the implementation in line with the declaration so
they actually match.

Closes #1322.