Fix test `Backtrace.comparison` for aggressive constant folding.
In order to test construction of backtraces we used a helper function
which was marked `noinline` to guarantee that it always constructed a
new stack frame. It seems this was not enough as a sufficiently smart
compiler might perform other optimizations like e.g., constant folding
which would also remove the stack frame.

We now disable optimizations for this function to make sure it is
emitted.

(cherry picked from commit 1333be6b40de6042d353c8b36f90bc80991e9442)