Remove performance pessimization introduced in `84fb4f21f494`.
When exposing this function directly we made the argument always
evaluate. This now causes us to compute strings for logging even if they
are never emitted.

Undo this change as it is actually not needed for the cleanup it was
initially intended for.

(cherry picked from commit d5fef37c379ac76762a7d8e2d9e3f9a417409a1b)