Fix pathological performance for `_haveEod` on highly chunked streams.
This function was identified as a performance bottleneck when parsing
`bytes` with a huge size from a stream with many chunks. The reason it
performed so poorly was due to its use of `unsafeEnd`: since `unsafeEnd`
creates an unsafe iterator from `end()` it always needs to compute the
chunk the safe iterator pointed to; if the stream contains many chunks
this causes the seen poor performance.

Since we only access `offset` use safe iterators instead. This makes it
much cheaper to call this function so we can safely use it during
parsing.

Since this change only affects performance we do not add a test case.
For a benchmark see #1345.

Closes #1345.