Fix incremental regexp matching for potentially empty results.
When feeding data incrementally into the regexp matcher, we could run
into an assert trying to ensure that the end of a match wasn't in the
previous chunk already. Turns out, however, that *is* possible with
empty matches that are determined only later (and maybe it's possible
in other cases, too). To fix this, we need to account for potentially
backtracking into data from a previous round. This could be a bit
surprising to callers of the `advance()` method, but should
work as long as they aren't trimming the stream in between (which in
our use cases shouldn't be the case).

Closes #1349.