Add DPD-style `spicy::accept_input()` and `spicy::decline_input()`.
This generalizes the accepting/decline input from being a Zeek thing to
being part of Spicy. Any Spicy grammar can call these functions as to
tell its host application if it recognizes the input format. The host
application can (but doesn't have to) install callbacks to receive
that information and react appropriately. For example, the Zeek
plugin will just forward it to Zeek's standard DPD hooks.

For testing, `spicy-driver` gets a new option `--require-accept`
that lets it exit with success only if `spicy::accept()` has
been called, and  `spicy::decline_input()` hasn't.

Note: The naming of the two functions is avoid conflicts with the
existing notion of synchronization's confirm/reject.