Search user module paths before system paths.
With, e.g., `spicyz -L /x/y/z`, we used to append the given module
path to the internal system paths, so that it was searched last. This
change turns around the priority to search `/x/y/x` first, before the
internal ones. That seems both more useful and more intuitive.