Use unique filename for all object files generated during JIT.
We previously would derive the name of the object file to generate for a
given C++ input file directly from the input filename by just replacing
the extension. This caused us to generate object files directly next to
C++ files added explicitly to the JIT job. Multiple jobs using the same
C++ input file could then race on the object file.

With this patch we make sure that object files are always generated with
taking the JIT has into account. We also fix the location of object
files to always be in a system temp directory.

Closes #1367.