Expose `barrier` type to Spicy.
We test it by sending bi-directional batch input into spicy-driver,
which we then parse with two units synchronizing on a joint barrier.

To make this work, `spicy-driver` learns to attempt resuming parsing
on the opposite side of a connection if that one currently blocked at
a barrier, in case the current data released that barrier.

The Zeek plugin will receive a similar change as `spicy-driver`.