Don't instantiate a debug logger if we aren't going to debug log.
This shortcuts the test in `HILTI_RT_DEBUG` to avoid a map lookup
every time when we already know we aren't going to log anything.

I was also wondering if `HILTI_RT_DEBUG` should be `#defined` to empty
if we're building a release version. However, being able to create
logs in release mode  seems quite helpful, and I believe with just a
pointer check now, the overhead should be minimal-—so not changing
that.