Fix stringification of enum values.
If an enum type or label used a C++ reserved identifier we would
previously stringify to name normalized for use in C++. Since we are
generating strings this is not needed and also confusing as stringified
names deviate from what the user wrote in the input.

With this patch we emit names which directly correspond to input
Spicy/HILTI names and labels.