Tune buffer memory management during stack switching.
We could end up doing a large amount of `reallocs()` during stack
switching when stack sizes kept changing. We now (1) enforce a minimum
size for the buffer where we save swapped out stack content, and (2)
do some rounding on what we allocate---both so that the buffer is more
likely to fit next time as well without needing reallocation. We also
switch from `realloc()` to `malloc()/free()` because we're doing a
subsequent `memcpy()` on our end anyways.

When measuring this with a specific trace, I actually didn't end up
seeing a substantial overall performance change. However, the memory
allocations did go down substantially as expected, and I can see the
old behavior degenerating quite a bit further, so I think this is
still a good change. The parameterization isn't necessarily final, we
can tune that more in the future (which still applies to all the fiber
options).