Don't store any parsed items for transient containers.
It's not quite as easy as it sounds. Turns out not all anonymous
containers should also be transient: With nested fields, the inner
fields will be anonymous, but they should still be storing values if
their top-level field is not.

This change is splitting out the notion of anonymous vs transient. The
former is strictly a function of whether we have an name for the
field. The latter is derived from the unit structure where the field
is located.

In the code most existing checks for being transient just become
checks for being anonymous (meaning, they don't change in their
semantics at all). The one difference is the piece we are aiming to
fix: we use the new, derived transient property to decide if a
container should be storing values.

The new behavior is tested through two existing tests:
`spicy/types/unit/transient.spicy` (w/ a couple of updates), and
`spicy/types/unit/nested-fields.spicy`.