Guard access to unit `_filters` member with feature flag.
The unit `_filters` member is accessed by most low-level functions
extracting data from the input, but its value is only interesting if a
unit actually supports filters. We had previously annotated `_filters`
with a requirement for filter support so any use of it with trigger
enabling support. By guarding access with a feature flag we can detect
such dependent uses in a subsequent patch.