Avoid unneeded temporaries when appending `View` to `Bytes`.
We used `View::data` to constructs a new `Bytes` to append from. Since
we cannot reuse the temporary in the append but have to copy the data
this introduced unneeded overhead. This patch inlines the low-level
`View` block iteration into `Bytes::append`.

I looked at this coming from zeek/spicy-dns#7, and with that particular
test case where many small views are appended the speedup is only ~1.5%.
I suspect the speedup to be bigger the bigger the view we append.

In order to amortize reallocation costs I also looked into aggressively
reserving capacity in the different `append` overloads, e.g., if the new
size is bigger than the current capacity, reserve double the needed
capacity in anticipation of more append operations to come. This did not
make a noticeable difference, probably due to the underlying
`std::string::append` already doing something similar.