Introduce runtime type for `self` values.
This introduces a runtime type `Self<T>` which can hold and access a raw
pointer without much overhead. It on purpose contains no safety net. The
intended use case is for `self` values where we know that for the scope
of `self` the pointed to value is always valid. As soon as we assign
`self` to some other value it needs to be turned into a managed
`ValueReference` again to manage safe access.

We will use this type in a subsequent patch.