Use `Self` to hold `self` values.
This patch changes the way `self` values are passed down into generated
parser functions. For that we change `ValueReference::self` to return a
`Self` and patch generated parser functions to take a `Self` instead of
a `ValueReference` as well. This is slightly hacky, but allows us to
punt on adding unsafe types to HILTI or Spicy since a `Self` represented
there could dangle; this is not an issue in generated code where we
always store values in managing `ValueReference` values which can be
constructed from `Self` values.

With this patch we can avoid performing safety checks on `self` at
runtime which depending on the parser should speed up performance (I saw
a speedup of 10% for a huge grammar I was benchmarking).